Revert "Nueva atencion por sede"
