+++ src/utils/axios.js
@@ -146,17 +146,14 @@ export const endpoints = {
     list: 'CalendarioController/getEspecialistaPorBeneficioYSede',
     modalities: 'CalendarioController/getModalidadesEspecialista',
     modalitiesBene: 'CalendarioController/getModalidadesEspecialistaBene',
-    area: 'Especialistas/area',
     sedes: 'Especialistas/sedes',
     horario: 'Especialistas/horario',
     horarios: 'Especialistas/horarios',
     disponibles: 'Especialistas/disponibles',
     meta: 'Especialistas/meta',
-    active: 'Especialistas/active',
   },
   areas: {
     citas: 'Areas/citas',
-    list: 'Areas/list',
   },
   calendario: {
     getAllEvents: 'CalendarioController/getAllEvents',
@@ -334,16 +331,6 @@ export const endpoints = {
     updateEstatusFaq: '/AyudaController/updateEstatusFaq',
     updateEstatusManual: '/AyudaController/updateEstatusManual',
   },
-  sedes: {
-    list: '/sedes/list',
-    save: '/sedes/save',
-    oficina: '/sedes/oficina',
-  },
-  oficinas: {
-    list: '/oficinas/list',
-  },
-  modalidades: {
-    list: '/modalidades/list',
   notificacion: {
     getNotificacion: '/NotificacionController/getNotificacion',
     deleteNotificacion: '/NotificacionController/deleteNotificacion'
